Add option score

+++ include/BoardMaster.h
@@ -23,7 +23,7 @@ public:
   void reverse_stone(int x, int y);
   void show();
   BoardMaster& operator=(BoardMaster& src);
-  int get_status_score();
+  double get_status_score();
 };
 
 char convert_stone_to_char(Stone src);
